# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UniChat API
  description: UniChat unified communication platform API
  version: 1.0.0
paths:
...
  /channels/{channel_id}/messages:
    get:
      tags:
      - messaging
      summary: List messages in a channel (cursor pagination)
      operationId: get_channel_messages_channels__channel_id__messages_get
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - messaging
      summary: Post a message to a channel
      operationId: create_message_channels__channel_id__messages_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /channels/{channel_id}/clear:
    post:
      tags:
      - messaging
      summary: Clear channel history for the current user only
      operationId: clear_channel_channels__channel_id__clear_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /messages/{message_id}/thread:
    get:
      tags:
      - messaging
      summary: Get thread replies for a message
      operationId: get_thread_messages__message_id__thread_get
      parameters:
      - name: message_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Message Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /messages/{message_id}:
    patch:
      tags:
      - messaging
      summary: Edit a message
      operationId: update_message_messages__message_id__patch
      parameters:
      - name: message_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Message Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessageRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - messaging
      summary: Soft delete a message
      operationId: delete_message_messages__message_id__delete
      parameters:
      - name: message_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Message Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClearChannelResponse:
      properties:
        channel_id:
          type: string
          format: uuid
          title: Channel Id
        cleared_at:
          type: string
          format: date-time
          title: Cleared At
      type: object
      required:
      - channel_id
      - cleared_at
      title: ClearChannelResponse
...
    CreateMessageRequest:
      properties:
        body:
          type: string
          maxLength: 4000
          minLength: 1
          title: Body
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
      type: object
      required:
      - body
      title: CreateMessageRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageAuthor:
      properties:
        id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Id
        name:
          type: string
          title: Name
        avatar_color:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Color
        is_external:
          type: boolean
          title: Is External
          default: false
      type: object
      required:
      - name
      title: MessageAuthor
    MessageResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        channel_id:
          type: string
          format: uuid
          title: Channel Id
        author:
          $ref: '#/components/schemas/MessageAuthor'
        body:
          type: string
          title: Body
        source:
          type: string
          title: Source
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        external_channel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Channel Id
        created_at:
          type: string
          format: date-time
          title: Created At
        edited_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Edited At
        is_deleted:
          type: boolean
          title: Is Deleted
          default: false
        reply_count:
          type: integer
          title: Reply Count
          default: 0
        last_reply_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Reply At
      type: object
      required:
      - id
      - channel_id
      - author
      - body
      - source
      - created_at
      title: MessageResponse
    MessagesPageResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      title: MessagesPageResponse
...
    ThreadResponse:
      properties:
        parent:
          $ref: '#/components/schemas/MessageResponse'
        replies:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Replies
      type: object
      required:
      - parent
      - replies
      title: ThreadResponse
    UpdateMessageRequest:
      properties:
        body:
          type: string
          maxLength: 4000
          minLength: 1
          title: Body
      type: object
      required:
      - body
      title: UpdateMessageRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...